CM-300 Error en reporte de animales vivos

+++ src/main/java/com/cema/administration/domain/report/Pregnancy.java
@@ -16,7 +16,7 @@ public class Pregnancy implements Reported<Pregnancy> {
     @ApiModelProperty(notes = "The year for this data", example = "2015")
     private Integer year;
     @ApiModelProperty(notes = "The percentage of pregnant cows", example = "72")
-    private Float percentage;
+    private Double percentage;
 
     @Override
     public int compareTo(@NotNull Pregnancy o) {
